+++ CHANGELOG.md
@@ -13,6 +13,10 @@ All notable changes to this project will be documented in this file.
 * Individual goal assessments (averaged to overall assessment)
 * Drill-down on goal analysis to view goal assessments
 * Show points on journal analysis chart for journals including selected hashtag or mention
+* Ability to edit tasks
+* Ability to edit habits in dashboard
+* Support up to 10 active habits (limit increased)
+* Sort habits by name
 
 ## Version 0-25